+++ CHANGELOG.md
@@ -2,6 +2,7 @@
 - Fix Missing Swamp Village Loot tables
 - Allow Trial Chambers Structures to spawn in/under BWG Biomes
 - Fix Incorrect English Language key for Granny Smith Advancement
+- Add the ability to disable Features added to Vanilla Biomes on NeoForge
 
 # 3.2.4
 - Fix Flat Plant Blocks having wood sounds, not Grass Sounds